PRECEDENCE | renamed DiscountCalculater to DiscountCalculator,  decomposed CalculateDiscount method.
